python: Enable library in static builds (for now)
pxr/external/boost/python was previously disabled if Python support
was explicitly disabled *or* if Python modules are disabled. The latter
happens in static builds regardless of whether Python support was
enabled or disabled by the user, which means this library was not
built under that configuration.

However, even without Python modules static builds still include code
in Tf and elsewhere that are Python-related. In order to update that
code to use headers from this library, we need to make sure it's built
even in static builds. We now disable this library only if Python
support is disabled AND Python modules are disabled.

The fact that we still build Python-related code in libraries even
when Python modules are disabled is surprising and probably a bug
to address later.

(Internal change: 2339975)