python: Fix crash when ref count of shared_ptr-held type drops to 0 in C++
This crash would occur because the shared_ptr_deleter that boost::python
attaches to shared_ptr-held types would call Python API without holding
the GIL. This change avoids that by ensuring the GIL is held in this case.

This is a known issue in upstream boost::python. It was discussed and
reported in a few different places:

https://mail.python.org/pipermail/cplusplus-sig/2007-June/012130.html
https://github.com/boostorg/python/pull/11

Pixar has been using this patch in internal builds since 2014. It was
never accepted upstream due to concerns about potential deadlocks
when a shared_ptr is being destroyed in one thread while another thread
is holding on to the GIL (see the GitHub PR above for more details).
While this may generally be true, Python-wrapped functions in OpenUSD
that invoke multi-threading typically ensure the GIL is dropped first
to avoid scenarios like the above.

This patch was originally written by John Loy (@jloy).

(Internal change: 2340607)