python: Remove boost::preprocessor use from clients requiring CV qualifiers
These clients used BOOST_PP_ENUM and BOOST_PP_ITERATE to create
template specializations for functions with different numbers of
arguments and different CV qualifiers. We now use variadic templates
to deal with the former, and a new PXR_BOOST_PYTHON_APPLY_QUALIFIERS
macro for the latter. This new macro takes some other macro and
calls it with each of the necessary CV qualifiers.

(Internal change: 2341259)