python: Remove boost::preprocessor use for invoking Python callables
These clients used boost::preprocessor to generate families of
functions that called PyObject_CallFunction or PyObject_CallMethod
with a varying number of arguments and a corresponding format string,
e.g. "(OOO)" when calling a function with 3 arguments.

The former is replaced with variadic templates. The latter is replaced
by calling PyObject_CallFunctionObjArgs and PyObject_CallMethodObjArgs.
These do the same thing as the previous functions but don't require
an extra format string as long as all arguments are given as PyObject*,
which they are in our use cases. As an extra bonus, per the Python C
API docs the new functions are faster than the previous ones.

(Note that for calls to PyObject_CallMethodObjArgs, we need to manually
convert the method name from char* const to a Python str via
PyUnicode_FromString. Per the cpython source code, this is exactly
what PyObject_CallMethod did anyway)

(Internal change: 2341412)