python: Add python::type_list to replace boost::mpl::vector
boost::python uses boost::mpl::vector to represent function
signatures, where the first element of the vector is the function's
return type and the subsequent elements are the function's argument
types. Clients also use this in the public API to specify signatures
for wrapped functions when they cannot be automatically deduced.

The new python::type_list is a simple type container that serves the
same purposes. Clients of mpl::vector throughout the library have been
updated to use this new type instead. python::type_list currently
derives from mpl::vector so it's still compatible with the remaining
mpl uses in the library, but this will be removed in a future change.

(Internal change: 2342640)