python: Remove boost::reference_wrapper and related utilities
This change adds python::reference_wrapper, ref, cref, which just
forward to the STL versions of these utilities. This change
also adds is_reference_wrapper and unwrap_reference, which require
manual implementations because the former has no STL equivalent
and the latter is not available in the STL until C++20.

(Internal change: 2342817)