python: Temporarily disable tests in static builds on Windows
These tests currently fail to build due to symbol export issues
on static builds. Tests in this configuration are nice to have
but not critical, since we don't build Python modules in static
builds anyway.

(If that's the case, you might wonder why we build this library
at all. See change 2339975 for answers.)

(Internal change: 2343231)